+++ src/main/kotlin/com/bibireden/playerex/api/PlayerEXTags.kt
@@ -1,0 +1,11 @@
+package com.bibireden.playerex.api
+
+import com.bibireden.playerex.PlayerEX
+import net.minecraft.core.registries.Registries
+import net.minecraft.tags.TagKey
+import net.minecraft.world.item.Item
+
+object PlayerEXTags {
+    @JvmField
+    val UNBREAKABLE_ITEMS: TagKey<Item> = TagKey.create(Registries.ITEM, PlayerEX.id("unbreakable"))
+}