+++ src/main/kotlin/com/bibireden/playerex/config/PlayerEXConfigModel.kt
@@ -54,6 +54,10 @@ class PlayerEXConfigModel {
 
     @Sync(SyncMode.OVERRIDE_CLIENT)
     @JvmField
+    var itemBreakingEnabled: Boolean = true
+
+    @Sync(SyncMode.OVERRIDE_CLIENT)
+    @JvmField
     var disableUI: Boolean = false
 
     @Sync(SyncMode.OVERRIDE_CLIENT)