+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## Changes 🌽
-- Reconciled connector based issue where old player data did not transfer over to a new player.
-- Updated the modded datapack to apply some attributes properly to scale.
+- Added "Item Breaking" mechanic
+- Added config option for Item Breaking